backend/pyproject.toml
[project]
name = "peer-review"
version = "0.1.0"
description = "Anonymous peer-review platform (FastAPI backend)"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "alembic>=1.14.0",
  "bcrypt>=4.2.0",
  "boto3>=1.34.0",
  "email-validator>=2.2.0",
  "fastapi>=0.115.6",
  "httpx>=0.28.1",
  "pdfplumber>=0.11.0",
  "psycopg[binary]>=3.2.3",
  "pydantic-settings>=2.6.1",
  "pytest>=9.0.2",
  "pywebpush>=2.0.0",
  "python-dotenv>=1.0.1",
  "python-jose[cryptography]>=3.3.0",
  "python-multipart>=0.0.17",
  "reportlab>=4.4.6",
  "ruff>=0.9.0",
  "sqlalchemy>=2.0.36",
  "ty>=0.0.9",
  "uvicorn[standard]>=0.32.1",
]

[tool.ruff]
exclude = [
    ".git",
    ".mypy_cache",
    ".ruff_cache",
    ".import_linter_cache",
    ".venv",
    "third_party",
]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
fixable = ["ALL"]
unfixable = []
select = [
    "A",  # flake8-builtin
    "B",  # flake8-bugbear
    "E",  # pycodestyle error
    "F",  # Pyflakes
    "I",  # isort
    "N",  # pep8-naming
    "W",  # pycodestyle warning
    "PL", # Pylint
    "UP", # pyupgrade
]
ignore = [
    "B905",    # Zip-without-explicit-strict
    "F403",    # Unable to detect undefined names
    "N812",    # Lowercase imported as non-lowercase
    "N999",    # Invalid module name
    "PLR0912", # Too many branches
    "PLR0913", # Too many arguments in function definition
    "PLR2004", # Magic value used in comparison
]

[tool.ruff.lint.isort]
force-single-line = true

[tool.ruff.format]
quote-style = "double"
line-ending = "auto"

[tool.importlinter]
root_packages = ["peer_review"]

[[tool.importlinter.contracts]]
name = "Layered architecture contract"
type = "layers"
layers = [
    "peer_review.infrastructure",
    "peer_review.adapters",
    "peer_review.use_cases",
    "peer_review.entities",
]
